Fix warnings in auth*, config-store*, core and cql

+++ cql/src/main/java/org/apache/cassandra/stargate/transport/internal/TransportDescriptor.java
@@ -38,7 +38,7 @@ public class TransportDescriptor {
     }
 
     public long toBytes(int val) {
-      return val * multiplier;
+      return (long) val * multiplier;
     }
   }
